Minor: move `SessionStateDefaults` into its own module
